auto-launcher same as Elden Ring, hate the goto tho
